Move to oyster directory.
