+++ oyster/stack_frame.py
