+++ oyster/symbol.py
